+++ cypress/integration/admin/professions/publish.spec.ts
@@ -1,3 +1,5 @@
+import { format } from 'date-fns';
+
 describe('Publishing professions', () => {
   context('When I am logged in as an editor', () => {
     beforeEach(() => {
@@ -18,6 +20,8 @@ describe('Publishing professions', () => {
       cy.translate('professions.admin.status.draft').then((status) => {
         cy.get('h2[data-status]').should('contain', status);
       });
+
+      cy.get('[data-cy=changed-by-user]').should('contain', '');
 
       cy.translate('professions.form.button.publishNow').then(
         (publishButton) => {
@@ -46,6 +50,19 @@ describe('Publishing professions', () => {
             cy.wrap($row).should('contain', status);
           });
         });
+
+      cy.contains('Gas Safe Engineer')
+        .parent('tr')
+        .within(() => {
+          cy.get('a').contains('View details').click();
+        });
+      cy.checkAccessibility();
+
+      cy.get('[data-cy=changed-by-user]').should('contain', 'Editor');
+      cy.get('[data-cy=last-modified]').should(
+        'contain',
+        format(new Date(), 'dd-MM-yyyy'),
+      );
     });
   });
 });