+++ src/professions/admin/profession-versions.controller.ts
@@ -15,18 +15,16 @@ import { I18nService } from 'nestjs-i18n';
 import { AuthenticationGuard } from '../../common/authentication.guard';
 import { BackLink } from '../../common/decorators/back-link.decorator';
 import { RequestWithAppSession } from '../../common/interfaces/request-with-app-session.interface';
-import { Legislation } from '../../legislations/legislation.entity';
 import { Nation } from '../../nations/nation';
 import { Organisation } from '../../organisations/organisation.entity';
 import QualificationPresenter from '../../qualifications/presenters/qualification.presenter';
-import { Qualification } from '../../qualifications/qualification.entity';
 import { UserPermission } from '../../users/user-permission';
 import { ShowTemplate } from '../interfaces/show-template.interface';
-import { ProfessionVersion } from '../profession-version.entity';
 import { Permissions } from '../../common/permissions.decorator';
 import { ProfessionVersionsService } from '../profession-versions.service';
 import { Profession } from '../profession.entity';
 import { ProfessionsService } from '../professions.service';
+import { ProfessionPresenter } from '../presenters/profession.presenter';
 
 @UseGuards(AuthenticationGuard)
 @Controller('/admin/professions')
@@ -71,6 +69,7 @@ export class ProfessionVersionsController {
     }
 
     const profession = Profession.withVersion(version.profession, version);
+    const presenter = new ProfessionPresenter(profession, this.i18nService);
 
     const organisation = Organisation.withLatestLiveVersion(
       profession.organisation,
@@ -94,6 +93,7 @@ export class ProfessionVersionsController {
 
     return {
       profession,
+      presenter,
       qualification: qualification,
       nations,
       industries,
@@ -113,34 +113,10 @@ export class ProfessionVersionsController {
         professionId,
       );
 
-    const newQualification = {
-      ...latestVersion.qualification,
-      id: undefined,
-      created_at: undefined,
-      updated_at: undefined,
-    } as Qualification;
-
-    const newLegislations = latestVersion.legislations.map((legislation) => {
-      return {
-        ...legislation,
-        id: undefined,
-        created_at: undefined,
-        updated_at: undefined,
-      } as Legislation;
-    });
-
-    const newVersion = {
-      ...latestVersion,
-      id: undefined,
-      status: undefined,
-      created_at: undefined,
-      updated_at: undefined,
-      user: req.appSession.user,
-      qualification: newQualification,
-      legislations: newLegislations,
-    } as ProfessionVersion;
-
-    const version = await this.professionVersionsService.save(newVersion);
+    const version = await this.professionVersionsService.create(
+      latestVersion,
+      req.appSession.user,
+    );
 
     return res.redirect(
       `/admin/professions/${version.profession.id}/versions/${version.id}/check-your-answers?edit=true`,
@@ -151,6 +127,7 @@ export class ProfessionVersionsController {
   @Permissions(UserPermission.PublishProfession)
   @Render('admin/professions/publish')
   async publish(
+    @Req() req: RequestWithAppSession,
     @Param('professionId') professionId: string,
     @Param('versionId') versionId: string,
   ): Promise<Profession> {
@@ -159,7 +136,12 @@ export class ProfessionVersionsController {
       versionId,
     );
 
-    await this.professionVersionsService.publish(version);
+    const newVersion = await this.professionVersionsService.create(
+      version,
+      req.appSession.user,
+    );
+
+    await this.professionVersionsService.publish(newVersion);
 
     return version.profession;
   }