+++ src/professions/presenters/profession.presenter.spec.ts
@@ -10,9 +10,12 @@ import { formatMultilineString } from '../../helpers/format-multiline-string.hel
 import { multilineOf } from '../../testutils/multiline-of';
 import { createMockI18nService } from '../../testutils/create-mock-i18n-service';
 import { translationOf } from '../../testutils/translation-of';
+import userFactory from '../../testutils/factories/user';
+import { formatDate } from '../../common/utils';
 
 jest.mock('../../nations/helpers/stringifyNations');
 jest.mock('../../helpers/format-multiline-string.helper');
+jest.mock('../../common/utils');
 
 describe('ProfessionPresenter', () => {
   let profession: Profession;
@@ -58,6 +61,56 @@ describe('ProfessionPresenter', () => {
 
       expect(formatMultilineString).toBeCalledWith(
         profession.qualification.level,
+      );
+    });
+  });
+
+  describe('changedBy', () => {
+    describe('when the Profession has been edited by a user', () => {
+      it('returns the name of the user', () => {
+        const profession = professionFactory.build({
+          changedByUser: userFactory.build({ name: 'beis-rpr' }),
+        });
+
+        const presenter = new ProfessionPresenter(
+          profession,
+          createMockI18nService(),
+        );
+
+        expect(presenter.changedBy).toEqual('beis-rpr');
+      });
+    });
+
+    describe("when the Profession hasn't yet been edited by a user", () => {
+      it('returns an empty string', () => {
+        const profession = professionFactory.build({
+          changedByUser: undefined,
+        });
+
+        const presenter = new ProfessionPresenter(
+          profession,
+          createMockI18nService(),
+        );
+
+        expect(presenter.changedBy).toEqual('');
+      });
+    });
+  });
+
+  describe('lastModified', () => {
+    it('should format the lastModified date on a profession', () => {
+      const profession = professionFactory.build({
+        lastModified: new Date('01-01-2022'),
+      });
+
+      const presenter = new ProfessionPresenter(
+        profession,
+        createMockI18nService(),
+      );
+
+      presenter.lastModified;
+      expect(formatDate as jest.Mock).toHaveBeenCalledWith(
+        new Date('01-01-2022'),
       );
     });
   });