+++ src/professions/profession.entity.spec.ts
@@ -24,6 +24,8 @@ describe('Profession', () => {
         qualification: professionVersion.qualification,
         reservedActivities: professionVersion.reservedActivities,
         legislations: professionVersion.legislations,
+        changedByUser: professionVersion.user,
+        lastModified: professionVersion.updated_at,
         status: professionVersion.status,
         versionId: professionVersion.id,
       });