+++ src/professions/profession.entity.ts
@@ -13,6 +13,7 @@ import { Industry } from '../industries/industry.entity';
 import { Legislation } from '../legislations/legislation.entity';
 import { Organisation } from '../organisations/organisation.entity';
 import { Qualification } from '../qualifications/qualification.entity';
+import { User } from '../users/user.entity';
 import {
   ProfessionVersion,
   ProfessionVersionStatus,
@@ -75,8 +76,10 @@ export class Profession {
   regulationUrl?: string;
   reservedActivities?: string;
   legislations?: Legislation[];
+  changedByUser?: User;
   versionId?: string;
   status?: string;
+  lastModified?: Date;
 
   constructor(
     name?: string,
@@ -161,6 +164,8 @@ export class Profession {
       protectedTitles: version.protectedTitles,
       regulationUrl: version.regulationUrl,
       legislations: version.legislations,
+      changedByUser: version.user,
+      lastModified: version.updated_at,
       status: version.status,
       versionId: version.id,
     };