+++ src/users/user.entity.ts
@@ -12,6 +12,7 @@ import {
 import { OrganisationVersion } from '../organisations/organisation-version.entity';
 import { Role } from './role';
 import { Organisation } from '../organisations/organisation.entity';
+import { ProfessionVersion } from '../professions/profession-version.entity';
 
 @Entity({ name: 'users' })
 export class User {
@@ -59,6 +60,12 @@ export class User {
     (organisationVersion) => organisationVersion.user,
   )
   organisationVersions: OrganisationVersion[];
+
+  @OneToMany(
+    () => ProfessionVersion,
+    (professionVersion) => professionVersion.user,
+  )
+  professionVersions: ProfessionVersion[];
 
   @ManyToOne(() => Organisation, (organisation) => organisation.users, {
     eager: true,