+++ CHANGELOG.md
@@ -11,6 +11,8 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 ### Changed
 
 - Users can enter a second legislation for a profession
+- Display user who changed a profession on the Admin Profession index page and Profession page itself.
+- Allow BEIS team users to see the "Changed By" field so they can track who has edited data.
 
 ## [release-003] - 2022-02-16